refactor: split extractor tag helpers
